powerpc/ci: Add ppce500 machine to kernel+qemu workflows
Recently we saw an issue with hugetlb max order folio pages.
The warning was getting reported on ppc64e but the problem also existed
on ppc32 for ppce500 machine type and e500mc cpu type for it's
unconventional max order of folio for hugetlbfs.

Hence enable kernel+qemu boot test with ppce500 machine as well to catch
such corner case issues in future.

Signed-off-by: Ritesh Harjani (IBM) <ritesh.list@gmail.com>